Move CallViewModel into ui-common module
