+++ app/src/main/kotlin/io/getstream/video/android/app/ui/call/CallActivity.kt
@@ -16,9 +16,9 @@
 
 package io.getstream.video.android.app.ui.call
 
+import io.getstream.video.android.common.viewmodel.CallViewModelFactory
 import io.getstream.video.android.compose.ui.AbstractComposeCallActivity
 import io.getstream.video.android.core.StreamVideo
-import io.getstream.video.android.core.viewmodel.CallViewModelFactory
 
 class CallActivity : AbstractComposeCallActivity() {
