+++ dogfooding/src/main/java/io/getstream/video/android/dogfooding/ui/call/CallActivity.kt
@@ -26,6 +26,8 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.size
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import io.getstream.video.android.common.viewmodel.CallViewModel
+import io.getstream.video.android.common.viewmodel.CallViewModelFactory
 import io.getstream.video.android.compose.theme.VideoTheme
 import io.getstream.video.android.compose.ui.components.call.CallContainer
 import io.getstream.video.android.compose.ui.components.call.controls.CallControls
@@ -37,8 +39,6 @@ import io.getstream.video.android.core.StreamVideo
 import io.getstream.video.android.core.call.state.ToggleCamera
 import io.getstream.video.android.core.call.state.ToggleMicrophone
 import io.getstream.video.android.core.permission.PermissionManager
-import io.getstream.video.android.core.viewmodel.CallViewModel
-import io.getstream.video.android.core.viewmodel.CallViewModelFactory
 import io.getstream.video.android.model.StreamCallId
 import io.getstream.video.android.model.streamCallId
