+++ stream-video-android-compose/src/main/kotlin/io/getstream/video/android/compose/ui/components/call/controls/CallControls.kt
@@ -29,13 +29,13 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import io.getstream.video.android.common.viewmodel.CallViewModel
 import io.getstream.video.android.compose.theme.VideoTheme
 import io.getstream.video.android.compose.ui.components.call.controls.actions.LandscapeCallControls
 import io.getstream.video.android.compose.ui.components.call.controls.actions.RegularCallControls
 import io.getstream.video.android.compose.ui.components.call.controls.actions.buildDefaultCallControlActions
 import io.getstream.video.android.core.call.state.CallAction
 import io.getstream.video.android.core.call.state.CallDeviceState
-import io.getstream.video.android.core.viewmodel.CallViewModel
 
 /**
  * Represents the set of controls the user can use to change their audio and video device state, or
