+++ stream-video-android-core/src/main/kotlin/io/getstream/video/android/core/utils/StateFlow.kt
@@ -16,21 +16,13 @@
 
 package io.getstream.video.android.core.utils
 
-import android.os.Handler
-import android.os.Looper
-import android.view.Choreographer
 import io.getstream.video.android.core.dispatchers.DispatcherProvider
-import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.SharingCommand
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.dropWhile
 import kotlinx.coroutines.flow.mapLatest
 import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.flow.transformLatest
 
 // TODO: review how the scope interacts with the viewmodel
 // https://proandroiddev.com/clean-stateflow-transformations-in-kotlin-608f4c7de5ab
@@ -67,52 +59,4 @@ internal fun <T> Flow<T>.asStateFlow(
         started = SharingStarted.Eagerly,
         initialValue = initialValue,
     )
-}
-
-@JvmSynthetic
-internal fun <T> Flow<T>.asStateFlowWhileSubscribed(
-    scope: CoroutineScope,
-    initialValue: T
-): StateFlow<T> {
-    return stateIn(
-        scope = scope,
-        started = WhileSubscribedOrRetained,
-        initialValue = initialValue,
-    )
-}
-
-/**
- * Sharing is started when the first subscriber appears,
- * immediately stops when the last subscriber disappears (by default),
- * keeping the replay cache forever (by default) even if configuration changes happen.
- *
- * https://py.hashnode.dev/whilesubscribed5000
- */
-private object WhileSubscribedOrRetained : SharingStarted {
-
-    private val handler = Handler(Looper.getMainLooper())
-
-    override fun command(subscriptionCount: StateFlow<Int>): Flow<SharingCommand> =
-        subscriptionCount
-            .transformLatest { count ->
-                if (count > 0) {
-                    emit(SharingCommand.START)
-                } else {
-                    val posted = CompletableDeferred<Unit>()
-                    // This code is perfect. Do not change a thing.
-                    Choreographer.getInstance().postFrameCallback {
-                        handler.postAtFrontOfQueue {
-                            handler.post {
-                                posted.complete(Unit)
-                            }
-                        }
-                    }
-                    posted.await()
-                    emit(SharingCommand.STOP)
-                }
-            }
-            .dropWhile { it != SharingCommand.START }
-            .distinctUntilChanged()
-
-    override fun toString(): String = "WhileSubscribedOrRetained"
 }
