+++ stream-video-android-ui-common/src/main/kotlin/io/getstream/video/android/common/viewmodel/CallViewModelFactory.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.getstream.video.android.core.viewmodel
+package io.getstream.video.android.common.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
