+++ stream-video-android-xml/src/main/kotlin/io/getstream/video/android/xml/binding/CallContainterViewBinding.kt
@@ -17,8 +17,8 @@
 package io.getstream.video.android.xml.binding
 
 import androidx.lifecycle.LifecycleOwner
+import io.getstream.video.android.common.viewmodel.CallViewModel
 import io.getstream.video.android.core.call.state.CallDeviceState
-import io.getstream.video.android.core.viewmodel.CallViewModel
 import io.getstream.video.android.xml.widget.callcontainer.CallContainerView
 import io.getstream.video.android.xml.widget.control.CallControlItem
